Splited TasksFeedPage into more pages
