+++ src/components/TasksFeedPage.js
@@ -1,17 +1,7 @@
 import "./TasksFeedPage.scss";
-import { useTasksFeed, CreateTask, FEED_SCOPE } from "../api/backend";
-import {
-    Switch,
-    Route,
-    Link,
-    Redirect,
-    useHistory
-} from "react-router-dom";
+import { useTasksFeed, FEED_SCOPE } from "../api/backend";
+import { Link } from "react-router-dom";
 import addIcon from "../svg/add-icon.svg";
-import TaskCreateComponent from "./TaskCreateComponent";
-import TaskPage from "./TaskPage";
-import moment from 'moment';
-import 'moment/locale/ru';
 import TaskComponent from "./TaskComponent";
 import SearchHeaderComponent from "./SearchHeaderComponent";
 import { useState } from "react";
@@ -24,56 +14,12 @@ const filters = [
 ]
 
 function TasksFeedPage() {
-    const history = useHistory();
-    moment.locale('ru')
-
     const profile = useSelector(state => state.profile);
 
     //TODO: handle errors
     const [scope, setScope] = useState(FEED_SCOPE.NOT_ASSIGNED);
     const [search, setSearch] = useState("");
     const feed = useTasksFeed(scope, search);
-
-    function OnCreateTask(input) {
-        if (!input.name)
-            return;
-
-        if (input.name.length > 128)
-            return;
-
-        if (!input.tags)
-            return;
-
-        if (input.tags.length < 1)
-            return;
-
-        if (input.tags.length > 5)
-            return;
-
-        if (!input.description)
-            return;
-
-        if (input.description.length > 2048)
-            return;
-
-        CreateTask({
-            name: input.name,
-            description: input.description,
-            tags: input.tags.map(tag => ({ ...tag, id: tag.new ? "MA==" : tag.id }))
-        }).then(function () {
-            history.push("/tasks");
-        }).catch(function (error) {
-            //TODO: handle errors
-            console.log(error);
-        });
-    }
-
-    function RenderTaskCreateComponent() {
-        if (profile.data?.customer)
-            return (<TaskCreateComponent onCreate={OnCreateTask} />)
-
-        return null
-    }
 
     function RenderTaskCreateButton() {
         if (profile.data?.customer)
@@ -101,31 +47,17 @@ function TasksFeedPage() {
         return null
     }
 
-    function RenderTaskPage() {
-        if (profile.loading || profile.data?.name)
-            return (<TaskPage />)
-        else return <Redirect to="/tasks" />
-    }
-
     return (
-        <Switch>
-            <Route exact path="/tasks">
-                <SearchHeaderComponent
-                    filters={filters}
-                    selectedFilter={scope}
-                    onFilterChange={value => setScope(value)}
-                    onInputChange={event => setSearch(event.target.value)}
-                />
-                {RenderTaskCreateButton()}
-                {RenderTaskList()}
-            </Route>
-            <Route path="/tasks/add">
-                {RenderTaskCreateComponent()}
-            </Route>
-            <Route path="/tasks/:id">
-                {RenderTaskPage()}
-            </Route>
-        </Switch>
+        <div className="flex-column overflow-auto">
+            <SearchHeaderComponent
+                filters={filters}
+                selectedFilter={scope}
+                onFilterChange={value => setScope(value)}
+                onInputChange={event => setSearch(event.target.value)}
+            />
+            {RenderTaskCreateButton()}
+            {RenderTaskList()}
+        </div>
     );
 }
